Poem-App/requirements.txt
psycopg2-binary
flask
flask-cors
openai
luma.core
luma.oled 
luma.lcd